Update cpp code

+++ c++/ilc.cpp
@@ -9,8 +9,8 @@ extern torch::Tensor RED(torch::Tensor);
 extern torch::Tensor ORTH(torch::Tensor);
 extern torch::Tensor CLP(int, torch::Tensor, torch::Tensor);
 
-const double mu_0 = 0.005;
-const double ratio = 200;
+const float mu_0 = 0.005;
+const float ratio = 200;
 const int T = 100000;
 const int T_r = 100;
 
@@ -22,16 +22,16 @@ torch::Tensor iterative_lattice_construction(int n) {
 
     for (int t = 0; t < T; ++t) {
         printf("%d/100000\r", t);
-        double mu = mu_0 * std::pow(ratio, -1.0 * t / (T - 1));
+        float mu = mu_0 * std::pow(ratio, -1.0 * t / (T - 1));
         torch::Tensor z = URAN(n);
         torch::Tensor y = z - CLP(n, B, torch::matmul(z, B));
         torch::Tensor e = torch::matmul(y, B);
 
         for (int i = 0; i < n; ++i) {
             for (int j = 0; j < i; ++j) {
-                B.index_put_({i, j}, B.index({i, j}) - mu * y[i].item<double>() * e[j].item<double>());
+                B.index_put_({i, j}, B.index({i, j}) - mu * y[i].item<float>() * e[j].item<float>());
             }
-            B.index_put_({i, i}, B.index({i, i}) - mu * (y[i].item<double>() * e[i].item<double>() - torch::norm(e).item<double>() * torch::norm(e).item<double>() / (n * B.index({i, i}).item<double>())));
+            B.index_put_({i, i}, B.index({i, i}) - mu * (y[i].item<float>() * e[i].item<float>() - torch::norm(e).item<float>() * torch::norm(e).item<float>() / (n * B.index({i, i}).item<float>())));
         }
         
         // printf("-- Current: Dim = %d, t = %d\n", n, t);
